Mark items that don't change at run-time as const
Those items that we don't expect to change during run-time
can be marked as const, so that they can be placed in 'read-only'
data segment. These items will then be on the flash, saving RAM.

One example of a common item that we don't expect to change dynamically
is the interface (that specifies methods and signals). Each element
in the array is a pointer (on 32-bit systems, each element takes up 4 bytes).

Change-Id: I753f81885938f23ef4aeb979e14575f50158b9b8