Fix build break on newer Ubuntu versions
TARG=linux (for async ReadFromStdIn) and TARG=yield-linux use pthread.
The newest Ubuntu version (13.04) seems to have disabled
implicit DSO linking. In earlier versions because we linked with librt,
it automatically pulled in its dependancy libpthread. It is no longer
the case in 13.04, and thus we need to explicitly call out our linking
dependancies.